Proxying vs "Rejected absolute url reference"
Experimental change to see if we pass tests when we first
check if the target fetch url is authorized with the domain
laywyer before we go on and check it with the extent to which
we are proxying according to the url namer.

Needs careful review considering the nature of the change.

Maybe fix https://github.com/pagespeed/mod_pagespeed/issues/1155